Moved rm in to the vcstools module and renmaed it rm_synth
