+++ vcstools/rm_synth.py
